+++ eurobot_camera/scripts/undistort.py
@@ -2,49 +2,24 @@
 
 import numpy as np
 import cv2
+
 from cv_bridge import CvBridge, CvBridgeError
 import rospy
-import sensor_msgs
 from sensor_msgs.msg import Image
-import sys, re
+from std_msgs.msg import String
+
 import yaml
-import matplotlib.pyplot as plt
-
 import copy
-
+import argparse
 import time
-import homogeneous
 
-# DIM=(2448, 2048)
-# K=np.array([[677.6729654637016, 0.0, 1235.371481335309], [0.0, 679.4736576804626, 1043.9887347932538], [0.0, 0.0, 1.0]])
-# D=np.array([[0.026055346132657364], [0.006035766757842894], [0.005666324884814231], [-0.0015661403498746557]])
-
-TABLE_LENGTH_Y=2.0
-TABLE_LENGTH_X=3.0
-TABLE_LENGTH_Z=1.0
-
-CAMERA_WIDTH=2448.
-CAMERA_HEIGHT=2048.
-HOMO_IMAGE_SCALE=2.
-HOMO_IMAGE_WIDTH=CAMERA_WIDTH/HOMO_IMAGE_SCALE
-HOMO_IMAGE_HEIGHT=CAMERA_HEIGHT/HOMO_IMAGE_SCALE
-
-CAMERA_X = 1.5
-CAMERA_Y = 0.0
-CAMERA_Z = 1.0
-CAMERA_ANGLE = (3 * np.pi / 4 - 0.1, 0, 0)
-
-PUCK_RADIUS=0.038
-pixel_scale_x=2448.0/3.0
-pixel_scale_y=2048.0/2.0
-PUCK_RADIUS_pixel_x=pixel_scale_x*PUCK_RADIUS
-PUCK_RADIUS_pixel_y=pixel_scale_y*PUCK_RADIUS
+from camera import Camera
 
 def read_config(conf_file):
     data_loaded = yaml.load(conf_file)
 
     DIM = (data_loaded['image_width'], data_loaded['image_height'])
-    print 'DIM:', DIM
+    rospy.loginfo("DIM:", DIM )
     camera_matrix = data_loaded['camera_matrix']
     data = camera_matrix['data']
     rows = camera_matrix['rows']
@@ -55,7 +30,7 @@ def read_config(conf_file):
         _K.append(data[i * rows:i * rows + cols:1])
 
     K = np.array(_K)
-    print 'camera matrix:', K
+    rospy.loginfo("CAMERA MATRIX:\n", K)
 
     distortion_coefficients = data_loaded['distortion_coefficients']
     data = distortion_coefficients['data']
@@ -67,273 +42,89 @@ def read_config(conf_file):
         _D.append(data[i * rows:i * rows + cols:1])
 
     D = np.array([[_D[0][0]], [_D[0][1]], [_D[0][2]], [_D[0][3]]])
-    print 'distortion coefficents:', D
+    rospy.loginfo("DISTORION COEFFICIENTS:\n", D)
 
     return DIM, K, D
-
-def find_rotation_matrix(theta):
-    R_x = np.array([[1, 0, 0],
-                    [0, np.cos(theta[0]),-np.sin(theta[0]) ],
-                    [0, np.sin(theta[0]), np.cos(theta[0])]])
-    R_y = np.array([[np.cos(theta[1]), 0, np.sin(theta[1])], 
-                    [0, 1, 0],
-                    [-np.sin(theta[1]), 0, np.cos(theta[1])]])
-    R_z = np.array([[np.cos(theta[2]), -np.sin(theta[2]), 0],
-                    [np.sin(theta[2]), np.cos(theta[2]), 0],
-                    [0, 0, 1]])
-    R = np.linalg.multi_dot([R_x, R_y, R_z])
-    return R
-
-class Camera():
-    def __init__(self, DIM, K, D):
-        self.DIM = DIM
-        self.K_camera = K
-        self.K_projection = K
-        self.D = D
-        self.warp_matrix = np.eye(3, 3, dtype=np.float32)
-        
-    def find_vertical_projection(self):
-        rotation_matrix = find_rotation_matrix(CAMERA_ANGLE)
-        camera_position = np.array([ [CAMERA_X],
-                                     [CAMERA_Y],
-                                     [CAMERA_Z] ])
-        t = -np.dot(rotation_matrix, camera_position)
-        # t = -rotation_matrix.dot(camera_position)
-        M = np.concatenate((rotation_matrix, t), axis=1)
-        L = np.array([ [TABLE_LENGTH_X/CAMERA_WIDTH, 0, 0],
-                       [0, -TABLE_LENGTH_Y/CAMERA_HEIGHT, TABLE_LENGTH_Y],
-                       [0, 0, 0],
-                       [0, 0, 1] ])
-        H = np.dot(M, L)
-        
-        K_new = np.linalg.inv(H)
-        
-        self.K_projection = K_new
-        
-        return K_new
-        
-    def filter_image(self, img):
-        return cv2.bilateralFilter(img,9,75,75)
-        
-    def find_thresholds(self, image_gray):
-        image = cv2.adaptiveThreshold(image_gray,255,cv2.ADAPTIVE_THRESH_GAUSSIAN_C,cv2.THRESH_BINARY,11,2)
-        return image
-        
-    def find_contours(self, image_gray):
-        ret, thresh = cv2.threshold(image_gray, 127, 255, 0)
-        image, contours, hierarchy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-        return image, contours
-        
-    def filter_contours(self, contours):
-        for cnt in contours:
-            #epsilon = 0.1*cv2.arcLength(cnt,True)
-            #approx = cv2.approxPolyDP(cnt,epsilon,True)
-            cnt = cv2.convexHull(cnt)
-        cnts = []
-        for cnt in contours:
-            area = cv2.contourArea(cnt)
-            perimeter = cv2.arcLength(cnt,True)
-            if area <= np.pi*PUCK_RADIUS_pixel_x*PUCK_RADIUS_pixel_y+500 and area >= np.pi*PUCK_RADIUS_pixel_x*PUCK_RADIUS_pixel_y-500 and perimeter <= 2*np.pi*PUCK_RADIUS_pixel_x+100 and perimeter >= 2*np.pi*PUCK_RADIUS_pixel_x-100 and len(cnt)>=5:
-                cnts.append(cnt)
-        return cnts
-        
-    def draw_contours(self, image, contours):
-        img = cv2.drawContours(image, contours, -1, (255,0,0), 3)
-        return img
-        
-    def watersherd(self, image):
-        gray = cv2.cvtColor(image,cv2.COLOR_BGR2GRAY)
-        ret, thresh = cv2.threshold(gray,0,255,cv2.THRESH_BINARY_INV+cv2.THRESH_OTSU)
-        # noise removal
-        kernel = np.ones((3,3),np.uint8)
-        opening = cv2.morphologyEx(thresh,cv2.MORPH_OPEN,kernel, iterations = 2)
-        # sure background area
-        sure_bg = cv2.dilate(opening,kernel,iterations=3)
-        # Finding sure foreground area
-        dist_transform = cv2.distanceTransform(opening,cv2.DIST_L2,5)
-        ret, sure_fg = cv2.threshold(dist_transform,0.7*dist_transform.max(),255,0)
-        # Finding unknown region
-        sure_fg = np.uint8(sure_fg)
-        unknown = cv2.subtract(sure_bg,sure_fg)
-        
-        # Marker labelling
-        ret, markers = cv2.connectedComponents(sure_fg)
-        # Add one to all labels so that sure background is not 0, but 1
-        markers = markers+1
-        # Now, mark the region of unknown with zero
-        markers[unknown==255] = 0
-        
-        markers = cv2.watershed(image,markers)
-        image[markers == -1] = [255,0,0]
-        cv2.imwrite("markers.jpg",markers)
-        return image
-        
-    def undistort(self, img):
-        undistorted_img = cv2.fisheye.undistortImage(distorted=img,
-                                                     K=self.K_camera,
-                                                     D=self.D,
-                                                     Knew=self.K_projection)
-        return undistorted_img
-    
-    def find_warp_matrix_not_feature(self, undistorted_img):
-        warp_matrix = homogeneous.homogeneous(undistorted_img,
-                                              HOMO_IMAGE_WIDTH,
-                                              HOMO_IMAGE_HEIGHT)
-        self.warp_matrix = warp_matrix
-        print warp_matrix
-        return warp_matrix
-    
-    def find_warp_matrix_feature(self, undistorted_img):
-        warp_matrix = homogeneous.alignImages(undistorted_img,
-                                              HOMO_IMAGE_WIDTH,
-                                              HOMO_IMAGE_HEIGHT)
-        self.warp_matrix = warp_matrix
-        print warp_matrix
-        return warp_matrix
-    
-    
-    def find_vertical_warp_projection(self, warp_matrix):
-        print warp_matrix
-        C = np.zeros((3,3))
-        C[(0,1,2),(0,1,2)] = HOMO_IMAGE_SCALE, HOMO_IMAGE_SCALE, 1
-        W = warp_matrix
-        C_inv = np.linalg.inv(C)
-        W_inv = np.linalg.inv(W)
-        K1 = C_inv.dot(self.K_projection)
-        K2 = W_inv.dot(K1)
-        K3 = C.dot(K2)
-        K_new = K3
-        self.K_projection = K_new
-        return K_new
 
 class CameraUndistortNode():
     def __init__(self, DIM, K, D):
         self.node = rospy.init_node('camera_undistort_node', anonymous=True)
-        self.publisher_undistorted = rospy.Publisher("/undistorted_image", Image)
-        self.publisher = rospy.Publisher("/recognition_image", Image)
-        self.publisher_gray = rospy.Publisher("/gray_scale_image", Image)
-        self.publisher_thresh = rospy.Publisher("/threshold_image", Image)
-        self.publisher_contours = rospy.Publisher("/contours_image", Image)
-        self.publisher_filter_contours = rospy.Publisher("/filtered_contours_image", Image)
-        self.subscriber = rospy.Subscriber("/usb_cam/image_raw", sensor_msgs.msg.Image, self.__callback)
+        self.publisher_undistorted = rospy.Publisher("/undistorted_image", Image, queue_size = 1)
+        self.publisher = rospy.Publisher("/recognition_image", Image, queue_size = 1)
+        self.publisher_gray = rospy.Publisher("/gray_scale_image", Image, queue_size = 1)
+        self.publisher_thresh = rospy.Publisher("/threshold_image", Image, queue_size = 1)
+        self.publisher_contours = rospy.Publisher("/contours_image", Image, queue_size = 1)
+        self.publisher_filter_contours = rospy.Publisher("/filtered_contours_image", Image, queue_size = 1)
+        self.publisher_coordinates = rospy.Publisher("/pucks_coordinates", String, queue_size = 1)
+        
         self.bridge = CvBridge()
         self.camera = Camera(DIM, K, D)
-        self.it = 0
         
-    
-    
-#     def horizontal_proj(self):
-#         rotation_matrix = self.__eulerAnglesToRotationMatrix((3 * np.pi / 4 - 0.1, 0, 0))
-#         camera_position = np.array([[1.5, 0.0, 1.0]]).T
-#         t = -rotation_matrix.dot(camera_position)
-#         print t
-#         M = np.concatenate((rotation_matrix, t), axis=1)
-#         L = np.array([[lx/hx, 0, 0], [0, 0, ly], [0, -lz/hy, lz], [0, 0, 1.0]])	
-#         K_new = np.linalg.inv(M.dot(L))
+        self.subscriber = rospy.Subscriber("/usb_cam/image_raw", Image,
+                                           self.__callback, queue_size = 1)
         
-#         return K_new
-
-
     def __callback(self, data):
-        try:
-            start_time = time.time()
-            cv_image = self.bridge.imgmsg_to_cv2(data, "bgr8")
-            rospy.loginfo(rospy.get_caller_id())
-        except CvBridgeError as e:
-            print(e)
+        start_time = time.time()
+        cv_image = self.bridge.imgmsg_to_cv2(data, "bgr8")
+        rospy.loginfo(rospy.get_caller_id())
         
+        # Process image
         cv_image = self.camera.filter_image(cv_image)
         undistorted_image = self.camera.undistort(cv_image)
         image = undistorted_image
-#         warp_matrix = self.camera.find_warp_matrix_feature(undistorted_image)
-#         self.camera.find_vertical_warp_projection(self.camera.warp_matrix)
-        if self.it == 0:
-            warp_matrix = self.camera.find_warp_matrix_feature(undistorted_image)
-            self.camera.find_vertical_warp_projection(self.camera.warp_matrix)
-            self.it += 1
-            return
 
-#         if self.it == 1:
-#             field = cv2.imread("/home/alexey/Desktop/field.png")
-#             field = cv2.resize(field, (2448,2048))
-#             image1 = undistorted_image*0.5+field*0.5
-#             cv2.imwrite("field_img_feature.jpg", image1)
-#             self.it += 1
-        try:
+        # Align image using field template
+        if self.camera.align_image(image):
+        
+            # Find thresholds and contours
             image_gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
-            #image_hsv = cv2.cvtColor(image, cv2.COLOR_RGB2HSV)
-            #image_gray = image_hsv[:,:,1] // 2 + image_hsv[:, :, 2] // 2
             image_thresholds = self.camera.find_thresholds(image_gray)
-            image_contours, contours = self.camera.find_contours(image_gray)
-            contours_image = copy.copy(image)
-            contours_image = self.camera.draw_contours(contours_image, contours)
-            cnts = self.camera.filter_contours(contours)
-            filter_contours_image = copy.copy(image)
-            filter_contours_image = self.camera.draw_contours(filter_contours_image, cnts)
-            image_ellipse = copy.copy(image)
-            image_ellipse_text = copy.copy(image)
-            for cnt in cnts:
-                ellipse = cv2.fitEllipse(cnt)
-                image_ellipse = cv2.ellipse(image_ellipse,ellipse,(0,255,0),2)
-                M = cv2.moments(cnt)
-                cx = M['m10']/M['m00']
-                cy = M['m01']/M['m00']
-                Cx = cx/pixel_scale_x
-                Cy = 2-cy/pixel_scale_y
-                
-                image_ellipse_text = cv2.putText(img=image_ellipse,
-                                                text="Cx=%.2f,Cy=%.2f" % (Cx, Cy),
-                                                org=(int(cx),int(cy)),
-                                                fontFace=cv2.FONT_HERSHEY_SIMPLEX,
-                                                fontScale=2,
-                                                color=(255,255,255),
-                                                thickness=5,
-                                                lineType=5)
-                print ("cx=", cx, " cy=", cy)
+            contours = self.camera.find_contours(image_gray)
+            image_contours = copy.copy(image)
+            image_contours = self.camera.draw_contours(image_contours, contours)
 
-                print ("Cx=", Cx, " Cy=", Cy)
-            #image = self.camera.draw_contours(image, contours)
-            field = cv2.imread("/home/alexey/Desktop/field.png")
-            field = cv2.resize(field, (2448,2048))
-            image1 = undistorted_image // 2 + field // 2
+            # Filter contours
+            contours_filtered = self.camera.filter_contours(contours)
+            image_filter_contours = copy.copy(image)
+            image_filter_contours = self.camera.draw_contours(image_filter_contours, contours_filtered)
+
+            # Create ellipse contours around pucks 
+            image_pucks = copy.copy(image)
+            coordinates = self.camera.find_pucks_coordinates(contours_filtered)
+            image_pucks = self.camera.draw_contours(contours_filtered, coordinates, image_pucks)
+
+            # Publish all images to topics
             self.publisher_undistorted.publish(self.bridge.cv2_to_imgmsg(image, "bgr8"))
-            self.publisher.publish(self.bridge.cv2_to_imgmsg(image_ellipse_text, "bgr8"))
-            self.publisher_gray.publish(self.bridge.cv2_to_imgmsg(image_gray)),# "bgr8"))
+            self.publisher_gray.publish(self.bridge.cv2_to_imgmsg(image_gray))
             self.publisher_thresh.publish(self.bridge.cv2_to_imgmsg(image_thresholds))
             self.publisher_contours.publish(self.bridge.cv2_to_imgmsg(contours_image, "bgr8"))
-            self.publisher_filter_contours.publish(self.bridge.cv2_to_imgmsg(filter_contours_image, "bgr8"))
-            res_time = time.time()-start_time
-            print (res_time)
-        except CvBridgeError as e:
-            print(e)
+            self.publisher_filter_contours.publish(self.bridge.cv2_to_imgmsg(image_filter_contours, "bgr8"))
+            self.publisher.publish(self.bridge.cv2_to_imgmsg(image_pucks, "bgr8"))
+            
+            # Publish pucks coordinates
+            self.publisher_coordinates.publish(coordinates)
 
+            res_time = time.time()-start_time
+            rospy.loginfo("RESULT TIME = ", res_time)
 
 if __name__ == '__main__':
-    conf_path_reg = re.compile(r'config_path=*')
-    conf_file = None
-    for arg in sys.argv[1:]:
-        if conf_path_reg.match(arg):
-            try:
-                conf_file = open(arg[12:], 'r')
-            except IOError as err:
-                sys.exit('Incorrect argument balance=')
-
-    if not conf_file:
-        print('Using default config file path')
-        try:
-            conf_file = open(
-                '/home/alexey/CatkinWorkspace/src/ros-eurobot-2019/eurobot_camera/configs/calibration.yaml', 'r')
-        except IOError as err:
-            sys.exit("Couldn't find config file")
-
-            
-    DIM, K, D = read_config(conf_file)
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-c", "--config",
+                        help="path to camera config yaml file",
+                        default="../configs/calibration.yaml")
+    parser.add_argument("-t", "--template",
+                        help="path to field's template file",
+                        default="../configs/field.png")
+    args = parser.parse_args()
 
     try:
-        undistort_node = CameraUndistortNode(DIM, K, D)
-        undistort_node.camera.find_vertical_projection()
-        rospy.spin()
-    except KeyboardInterrupt:
-        print("Shutting down")
-    cv2.destroyAllWindows()
+        conf_file = open(args.config, 'r')
+    except IOError as err:
+        sys.exit("Couldn't find config file")
+        
+    DIM, K, D = read_config(conf_file)
 
+    undistort_node = CameraUndistortNode(DIM, K, D)
+    undistort_node.camera.find_vertical_projection()
+    
+    rospy.spin()
